ml-service/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Machine Learning Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
statsmodels==0.14.0
arch==6.2.0

# Data Processing
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2

# API and Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2

# Logging and Monitoring
structlog==23.1.0

# Development Dependencies
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Optional: For advanced ML features
# xgboost==1.7.6
# lightgbm==4.0.0
# plotly==5.15.0